third_party/WebKit/LayoutTests/inspector/elements/selected-element-changes-execution-context-expected.txt

Tests that the execution context is changed to match new selected node.


Context has changed:
true
Context matches node frameId: 
true
Context has changed:
true
Context matches node frameId: 
true
